Finished moving specializations of edge builder to a new header only included through the source of planning graph.
